updated response type

+++ src/main/java/com/taxicalls/notifications/repository/NotificationRepository.java
@@ -3,6 +3,6 @@ package com.taxicalls.notifications.repository;
 import com.taxicalls.notifications.model.Notification;
 import org.springframework.data.repository.CrudRepository;
 
-public interface NotificationRepository extends CrudRepository<Notification, Integer> {
+public interface NotificationRepository extends CrudRepository<Notification, Long> {
 
 }
